% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liladata_class.R
\name{as.data.frame.liladata}
\alias{as.data.frame.liladata}
\title{Umwandeln von \code{liladata} in \code{data.frame}. Dabei werden alle Metinformationen mit den Werten zusammengeführt.}
\usage{
\method{as.data.frame}{liladata}(.data)
}
\arguments{
\item{.data}{Objekt mit Klasse \code{liladata}}
}
\value{
Objekt mit Klasse \code{data.frame}
}
\description{
Umwandeln von \code{liladata} in \code{data.frame}. Dabei werden alle Metinformationen mit den Werten zusammengeführt.
}
\examples{
data <- read_lila_block(system.file("extdata", "exp_lilablock.lila", package="lilatools"))
as.data.frame(data)
}
\seealso{
data.frame

Other liladata_methods: \code{\link{liladata}},
  \code{\link{merge_liladata}},
  \code{\link{select.liladata}}
}
\concept{liladata_methods}
