% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reading.R
\name{check_legformat}
\alias{check_legformat}
\title{Dateinamen im LEG-Format pr체fen und zerlegen}
\usage{
check_legformat(filename)
}
\arguments{
\item{filename}{zu pr체fender Dateiname (character)}
}
\value{
\code{list} mit den Elementen \code{check} (logische Angabe ob file dem Format entspricht) und \code{leg_format} (Informationen im Dateinamen als \code{data.frame})
}
\description{
Die Konformit채t des Dateinamens mit den Vorgaben zum LEG-Format werden gepr체ft. Die einzelnen Informationen im Dateinamen werden extrahiert
}
\examples{
check_legformat("BY_24215903-tzugtsvhs-e13-201805150000-LARSIM_WHM-201805100500-VAR1.lila")
check_legformat("BY_24215903-qvhs.lila")
}
