% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liladata_class.R
\name{liladata}
\alias{liladata}
\title{Erstellt ein Objekt der Klasse \code{liladata}}
\usage{
liladata(data, meta)
}
\arguments{
\item{data}{data.frame enthält die tatsächlichen Werte des Datensatzes}

\item{meta}{data.frame mit den Metainformationen}
}
\value{
Objekt der Klasse liladata
}
\description{
Die Klasse ermöglicht das Verarbeiten Daten aus Dateien im LILA-Format. Insbesondere werden die gesamten Metainformationen in der Klasse beibehalten.
}
\details{
Die Lese-Funktionen von LILA-Dateien erstellen ein Objekt der S3-Klasse \code{liladata}. Dieses enthält alle aus den LILA-Dateien ausgelesenen Daten. Für die Weiterverarbeitung kann das Objekt mit \code{\link[as.data.frame.liladata]{as.data.frame}} in ein data.frame umgewandelt werden. Dieser Weg ist aber nicht empfohlen, da LILA-Dateien üblicherweise umfangreiche Meta-Informationen haben, die für die gewünschte Analyse nicht erforderlich sind. Daher sollte zunächst eine Auswahl der erfordelrichen Daten mit \code{\link{select.liladata}} erfolgen.

Intern trennt die Klasse die Daten der Zeitreihen von den Metainformationen. Dadurch soll der Speicherbedarf minimiert werden, was bei großen Datensätzen relevant werden kann. Daher enthält die Klasse zwei Listenelemente mit den Namen: data und meta. Diese können, wie von Listen gewohnt über '$data' oder '$meta' aufgerufen werden.
Die Daten im data-Listenelement sind als data.frame oder tibble abgelegt mit den Spalten:
\describe{
\item{time}{die Zeitschritte,}
\item{id}{eine eindeutige Kennzeichnung eines jeden Werteblocks und}
\item{value}{die jeweiligen Werte}
}
Die Metainformationen  im meta-Listenelement sind ebenfalls als data.frame oder tibble abgelegt. Dort ist ebenfalls die id enthalten, mit der die Metainformationen den Werteblöcken eindeutig zugeordnet werden kann.
Als weitere Informationen müssen mindestens die folgenden Spalten enthalten sein, wie dies für das LILA-Format definiert ist:
\itemize{
\item station
\item datenart
\item dimension
\item zeitintervall
}
Zusätzlich sind beliebig viele weitere Spalten möglich.

Eine Veränderung der id ist nicht zu empfehlen, und sollte wenn, dann sowohl in \code{data} als auch in \code{meta} erfolgen. Normalerweise ist eine Manipulation auch nicht erforderlich. Jegliche Veränderung und Veränderung der Daten in R sollte erst nach einer Auswahl und Umwandlung in ein data.frame oder tibble erfolgen. Dies macht die Analyse ohnehin einfacher.
}
\examples{
# empty object
liladata()

# example object
liladata(data=data.frame(id="1-1",
                         time=c("09.09.2019 00:00", "09.09.2019 01:00", "09.09.2019 02:00",
                                "09.09.2019 03:00", "09.09.2019 04:00", "09.09.2019 00:50"),
                         values=c(0.9, 0.9, 0.9, 0.9, 0.8, 0.8)),
         meta=data.frame(id='1-1', station="Rebbelroth", gewaesser="Agger",
                         stationsnummer=01320010, stationskennung='REBR',
                         datenart='Q',dimension='cbm/s', datenbezug='GTS',
                         zeitintervall='01:00', datentyp='M',
                         datenursprung='vhs', flaeche=109.860, flusskilometer=63.693,
                         vorhersagezeitpunkt='08.09.2019 13:00',
                         kommentar='Abfluss Vorhersage',
                         austauschkennung='RP_01320010-qvhs-d01-201908270700'
                         )
                         )
}
\seealso{
Other liladata_methods: \code{\link{as.data.frame.liladata}},
  \code{\link{merge_liladata}},
  \code{\link{select.liladata}}
}
\concept{liladata_methods}
