% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liladata_class.R
\name{merge_liladata}
\alias{merge_liladata}
\title{Zusammenfügen zweier \code{liladata}-Objekte}
\usage{
merge_liladata(x, y)
}
\arguments{
\item{x}{Objekt mit Klasse \code{liladata}}

\item{y}{Objekt mit Klasse \code{liladata}}
}
\value{
Objekt mit Klasse \code{liladata}
}
\description{
Zwei Objekte mit Klasse \code{liladata} werden zusammengeführt. Dabei werden jeweils die Daten- und Metainformationen vereint. Dazu werden die ids angepasst, um Überschneidungen zwischen den beiden Objekten zu verhindern.
}
\details{
Im Zuge der Zusammenführung werden die IDs der beiden Objekte zusammengeführt. Dabei wird verhindert, dass Überschneidungen auftreten. Die ID ist also unter Umständen nicht mehr mit der ID der originalen Datensätze identisch.
}
\examples{
x <- read_lila_block(system.file("extdata", "exp_lilablock.lila", package="lilatools"))
y <- read_lila_einzel(system.file("extdata", "exp_lilaeinzel.lila", package="lilatools"))
merge_liladata(x, y)
}
\seealso{
Other liladata_methods: \code{\link{as.data.frame.liladata}},
  \code{\link{liladata}}, \code{\link{select.liladata}}
}
\concept{liladata_methods}
