% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reading.R
\name{read_lila}
\alias{read_lila}
\title{Funktion zum Einlesen von Dateien im LILA format}
\usage{
read_lila(file, lila_einzel = FALSE, ...)
}
\arguments{
\item{file}{path of file (char) to read}

\item{lila_einzel}{hat einzulesende Datei das Format LILA-EINZEL (optional)}

\item{...}{andere Optionen, siehe Details}
}
\value{
Objekt mit Klasse \code{\link{liladata}}
}
\description{
Die Funktion ermöglicht das Einlesen aller Varianten des LILA-Formats
}
\details{
LILA-Dateien in den LILA-Formaten SPALTE, EINZEL, BLOCK und hybrid können eingelesen werden. Für LILA-EINZEL ist eine vereinfachte und daher beschleunigte Routine implementiert, die über das Argument \code{lila_einzel=TRUE} aktiviert werden kann.

Wenn Dateien im Format LILA EINZEL eingelesen werden, kann die Option \code{filename_metainfo=TRUE} gesetzt werden, der Dateiname dem LEG-Format entspricht. DAnn werden die im Dateinamen enthaltenen Informationen den Metainformationen hinzugefügt.
}
\examples{
read_lila(system.file("extdata", "exp_lilaeinzel.lila", package="lilatools"))
read_lila(system.file("extdata", "exp_lilaeinzel.lila", package="lilatools"), lila_einzel=T)
read_lila(system.file("extdata", "exp_lilablock.lila", package="lilatools"))
}
\seealso{
Other read_lila: \code{\link{read_lila_block}},
  \code{\link{read_lila_einzel}}
}
\concept{read_lila}
