% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reading.R
\name{read_lila_block}
\alias{read_lila_block}
\title{Einlesen von LILA-BLOCK-Format}
\usage{
read_lila_block(file, filename_metainfo = F, locale = "default",
  n_max = Inf)
}
\arguments{
\item{file}{Pfad zur einzuleseneden Datei}

\item{filename_metainfo}{(logisch, default: FALSE) enthält der Dateipfad Metadaten, die importiert werden sollen?}

\item{locale}{setzt einige wichtige Informationen für den Lesevorgang, default siehe Details}

\item{n_max}{maximale Anzahl an Zeilen, die aus dem Datenblock (also unterhalb der Metadaten) eingelesen werden}
}
\value{
Objekt in liladata-Klasse
}
\description{
Die Funktion ist eine flexible Routine, um eine Datei im LILA-BLOCK-Format einzulesen.
}
\details{
\code{read_lila_block} ist die flexibelste der \code{read_lila}-Funktionen: Sie kann neben dem Format LILA-BLOCK auch LILA EINZEL, LILA SPALTE und LILA-hybrid einlesen. Dabei bestehen aber Geschwindigkeitsnachteile.

\code{locale} hat den folgenden default-Wert:

locale(decimal_mark = ".", date_names = "de", encoding="latin1", time_format="\%d.\%m.\%Y \%H:\%M")

Er kann verändert werden, wozu indem der angegebene Aufruf von locale kopiert wird und nach bedarf verändert wird. Dabei sollten die mindestens die verwendeten Parameter gesetzt werden.
Es empfiehlt sich, \code{locale} mit den Standardparametern einmal für eine Session zu definieren, besonders wenn viele Dateien eingelesen werden sollen. Locale hat leider einen relativ großen Rechenbedarf beim Aufruf.
Wenn es nur einmal aufgerufen wird hat das also große Geschwindigkeitsvorteile (siehe Beispiel).
}
\examples{
# read_lila_block(system.file("extdata", "exp_lilablock.lila", package="lilatools"))
# locale extern aufrufen
loc <- locale(decimal_mark = ".", date_names = "de", encoding="latin1")
read_lila_block(system.file("extdata", "exp_lilablock.lila", package="lilatools"), locale=loc)
}
\seealso{
Other read_lila: \code{\link{read_lila_einzel}},
  \code{\link{read_lila}}
}
\concept{read_lila}
