% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reading.R
\name{read_lila_einzel}
\alias{read_lila_einzel}
\title{Lesen von Dateien im LILA-EINZELDATEI-Format}
\usage{
read_lila_einzel(file, filename_metainfo = FALSE, locale = "default",
  n_max = Inf)
}
\arguments{
\item{file}{Pfad der zu lesenden Datei}

\item{filename_metainfo}{logisch (default: FALSE), inspiziert Dateinamen mit Standardformat nach zusätzlichen Metainformationen (siehe Details)}

\item{locale}{Lokaleinstellungen, können extern vorgegeben werden, default: locale(decimal_mark = ".", date_names = "de", encoding="latin1")}

\item{n_max}{maximale Anzahl an Zeilen, die aus dem Datenblock (also unterhalb der Metadaten) eingelesen werden}
}
\value{
Objekt der Klasse \code{\link{liladata}}
}
\description{
Die Funktion ist optimiert fürs Einlesen von Dateien im LILA-EINZELDATEI-Format
}
\details{
Eine Datei im LARSIM format LILA EINZELDATEI wird eingelesen. Abweichende LILA-Formate können nicht gelesen werden.
Wenn der der Dateiname der 'Formatdefinition der Austauschkennung zum länderübergreifenden Austausch von Vorhersagen V 1.0' (LEG) entsprcht, werden bei angabe der der Option inspect_filename zusätzliche Informationen aus dem Dateinamen entnommen.

Die Lokaleinstellungen mit \code{locale} umfassen zB die Einstellungen für das Dateiencoding. In der Regel muss daran nichts geändert werden. Allerdings empfiehlt es sich das locale extern als Variable zu speichern und vorzugeben, wenn viele Dateien eingelesen werden. Dann wird das Einlesen erheblich beschleunigt.
}
\examples{
read_lila_einzel(system.file("extdata", "exp_lilaeinzel.lila", package="lilatools"))
}
\seealso{
Other read_lila: \code{\link{read_lila_block}},
  \code{\link{read_lila}}
}
\concept{read_lila}
