% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liladata_class.R
\name{select.liladata}
\alias{select.liladata}
\title{Datenauswahl erstellen}
\usage{
\method{select}{liladata}(.data, ...)
}
\arguments{
\item{.data}{Objekt der Klasse \code{liladata}}

\item{...}{Auswahl der gewünschten Spalten, getrennt durch Kommas}
}
\value{
data.frame mit den Spalten id, time, values ad_lilaund den gewählten Spalten mit Metainformationen.
}
\description{
Die gewüschten Metainformationen werden aus einem \code{liladata}-Objekt ausgewählt, mit den Werten zusammengeführt und in ein \code{data.frame} umgewandelt.
}
\details{
Die Auswahl der gewünschten Spalten erfolgt bequem wie bei der Funktion \code{\link[dplyr]{select}}. Die Spaltennamen können direkt, durch Kommas getrennt, aber ohne Anführungsstriche aufgeführt werden. Alternativ kann vor die Spaltennamen auch ein Minus '-' gestellt werden. Dann werden die betroffenen Spalten nicht verwendet.

Auch weitere Auswahlfunktionen können verwendet werden (siehe dazu die Dokumentation von \code{\link[dplyr]{select}})

Eine Sonderstellung nimmt die id-Spalte ein. Diese wird immer ausgegeben, da sie für das Zusammenführen von Werten und Metainformationen verwendet wird. Innerhalb der Klasse liladata stellt sie sicher, dass Metadaten und Werte korrekt zugeordnet werden. Wenn sie nicht mehr benötigt wird, kann die id-Spalte also gelöscht werden. Im Fall, dass die Daten aber in das ursprüngliche \code{liladata}-Objekt zurückgeführt werden soll, sollte die id-Spalte erhalten bleiben.

Bei der Auswahl sollte darauf geachtet werden, das ausreichend Metainformationen ausgegeben werden, sodass alle Werte eindeutig identifizert werden. Wenn also beispielsweise zwei Stationen denselben Namen tragen, sollte stattdessen eine eindeutig definierte Spalte verwendet werden, zB "Stationskennung".
}
\seealso{
\code{\link{liladata}} für Informationen zur \code{\link{liladata}}-S3-Klasse, \code{\link[dplyr]{select}} für weitere Informationen über Auswahl-Funktionalitäten

Other liladata_methods: \code{\link{as.data.frame.liladata}},
  \code{\link{liladata}}, \code{\link{merge_liladata}}
}
\concept{liladata_methods}
